==> code/1_convert_hetionet_nodes_to_umls.ipynb <==
# Convert Hetionet nodes to UMLS space

## Workflow

1. Run Daniel's `integrate.ipynb` to get `data/hetnet.json.bz2` file. All edges from PharmacotherapyDB were excluded.
2. Run `prepare_neo4j_import_csvs.ipynb` to convert to CSV.
3. Map all ids to UMLS CUIs whenever possible.

import pandas as pd
import re

from collections import defaultdict

## Read Hetionet nodes

nodes = (pd
    .read_csv("../data/hetionet/raw/hetionet_nodes.csv", sep=',')
    .drop("identifier:string", axis=1)
    .rename(columns={
        ":ID": "hetio_id",
        "name:string": "name",
        ":LABEL": "het_type"
    })
)

nodes.shape

nodes.head()

---

## Create unified ID map

id_map = defaultdict(list)

def add_to_map(df, start, stop):
    id_map["orig"].extend(list(map(str, df[start])))
    id_map["dest"].extend(df[stop])

---

(pd
    .read_csv("../pipeline/maps/uberon.tsv", sep='\t')
    .pipe(add_to_map, "uberon_id", "cui")
)

(pd
    .read_csv("../pipeline/maps/drugbank.tsv", sep='\t')
    .pipe(add_to_map, "drugbank_id", "cui")
)

(pd
    .read_csv("../pipeline/maps/doid_to_cui.tsv", sep='\t')
    .pipe(add_to_map, "doid", "cui")
)

(pd
    .read_csv("../pipeline/maps/gene_map.tsv", sep='\t')
    .pipe(add_to_map, "entrez_id", "cui")
)

(pd
    .read_csv("../pipeline/maps/remaining.tsv", sep='\t')
    .pipe(add_to_map, "code", "cui")
)

---

id_map = pd.DataFrame(id_map)

id_map.shape

id_map.head()

---

### Data validation

id_map["dest"].str.startswith("UMLS:C").value_counts()

## Perform mapping

def is_cui(s):
    return re.match(r'^C[0-9]{7}$', s) is not None

hnodes = (nodes
    .merge(id_map, how="left", left_on="hetio_id", right_on="orig")
    .drop("orig", axis=1)
        
    .assign(val = lambda df: df["dest"].fillna(df["hetio_id"]))
    .drop("dest", axis=1)
    
    .assign(
        cui = lambda df: df["val"].map(
            lambda v: "UMLS:{}".format(v) if is_cui(v) else v
        )
    )
    .drop("val", axis=1)
          
    .sort_values(["hetio_id", "cui", "name", "het_type"])
    .reset_index(drop=True)
)

hnodes.shape

hnodes.head()

## Save to file

hnodes.to_csv("../pipeline/hetionet/hetionet_nodes_umls.tsv", sep='\t', index=False)
