==> code/2_determine_how_to_merge_networks.ipynb <==
# Determine how to merge Hetionet with SemmedDB

Executive decision after discussion: overwrite hetio types with semmed types. For any remaining hetio types without cuis in semmed, just change the type to the most common one

Keep any hetio nodes with missing cuis in hetio id space

**new:**

Try to merge semmeddb types and edges into hetionet space. this reduces the complexity, and allows us to try running rephetio on the resulting network

import pandas as pd

## Read Hetionet nodes

hnodes = pd.read_csv("../pipeline/hetionet/hetionet_nodes_umls.tsv", sep='\t')

hnodes.shape

hnodes.head()

---

## Read SemmedDB nodes

snodes = (pd
    .read_csv("../data/semmeddb/no_treats/semmeddb_no_treats_nodes.tsv", sep='\t')
    .rename(columns={
        "node_type": "sem_type",
        "node_id": "cui"
    })
)

snodes.shape

snodes.head()

---

## Determine the most common SemmedDB node type for each Hetionet node type

ntype_map = (hnodes
    .drop("name", axis=1)
    .merge(snodes, how="inner", on="cui")
    .drop("node_name", axis=1)

    .groupby("het_type")
    .apply(lambda df: df["sem_type"].value_counts().axes[0][0])
    .to_frame("sem_type")
    .reset_index()
             
    .append(
        pd.DataFrame(
            [("Pathway", "Pathway")],
            columns=["het_type", "sem_type"]
        )
    )
    .reset_index(drop=True)
)

ntype_map

Hetionet nodes which have CUIs that exist in SemmedDB will just get mapped to that CUI.

Hetionet nodes which do not exist in SemmedDB will use their original Hetionet id. Their node type is converted to the node type most commonly associated with that Hetionet node type (through direct CUI mappings).

---

## Determine how Hetionet nodes will be identified in the merged network

Rules:
* If the Hetionet node's CUIs are not in SemmedDB: use the original Hetionet id
* If the Hetionet node has N CUIs, then only retain those which exist in SemmedDB (discard the non-existent ones)

all_scuis = set(snodes["cui"])

hnode_map = (hnodes
    .assign(in_sem = lambda df: df["cui"].map(lambda v: v in all_scuis))
    .groupby("hetio_id")
    .apply(
        lambda df: df.query("in_sem") if df["in_sem"].any() else df.head(1)
    )
    .reset_index(drop=True)
    .assign(final_id = lambda df:
        df[["hetio_id", "cui", "in_sem"]].apply(
            lambda r: r["cui"] if r["in_sem"] else r["hetio_id"],
            axis=1
        )
    )
    .drop("cui", axis=1)
)

hnode_map.shape

hnode_map.head()

---

### Resulting stats

hnode_map["hetio_id"].nunique()

hnode_map["final_id"].nunique()

### How many unique Hetionet nodes were mapped?

hnode_map.groupby("in_sem")["hetio_id"].nunique()

### Of the unmappable Hetionet nodes, what types were they?

hnode_map.query("~in_sem")["het_type"].value_counts()

### How many identifiers per hetio id (as a %)

(hnode_map
    .groupby("in_sem")
    .apply(
        lambda df: (df
            .groupby("hetio_id")
            .size()
            .value_counts(normalize=True)
            .multiply(100)
        )
    )
)

Of the Hetionet nodes which could be mapped to UMLS CUIs, the vast majority (97%) mapped to a single CUI.

---

## Change Hetionet semantic types to SemmedDB semantic types

fin_hnode_map = (hnode_map
    .merge(
        snodes.drop("node_name", axis=1), how="left",
        left_on="final_id", right_on="cui"
    )
    .drop("cui", axis=1)
    .rename(columns={"sem_type": "orig_stype"})
    .merge(ntype_map, how="left", on="het_type")
    .assign(
        fin_type = lambda df: df["orig_stype"].fillna(df["sem_type"])
    )
    .drop(["orig_stype", "sem_type"], axis=1)
    .reset_index(drop=True)
)

fin_hnode_map.shape

fin_hnode_map.head()

## Save to disk

fin_hnode_map.to_csv(
    "../pipeline/hetionet/final_hetionet_nodes_umls.tsv", sep='\t', index=False
)

---

# Determine how to map Hetionet edges

Need to change Hetionet edges to the edge types in SemmedDB.

## Map Hetionet edges to SemmedDB edges

Adjust the semantics of the edges accordingly, and simplify where possible.

hedges = (pd
    .read_csv("../data/hetionet/raw/hetionet_edges.csv", sep=',')
    .rename(columns={
        ":START_ID": "source_id",
        ":END_ID": "target_id",
        ":TYPE": "het_etype"
    })
)

hedges.shape

hedges.head()

---

### Convert Hetionet edges to their final node types

het_etypes = (hedges
    .merge(
        fin_hnode_map[["hetio_id", "het_type", "fin_type"]],
        how="inner", left_on="source_id", right_on="hetio_id"
    )
    .drop("hetio_id", axis=1)
    .rename(columns={
        "het_type": "source_htype",
        "fin_type": "source_ftype"
    })

    .merge(
        fin_hnode_map[["hetio_id", "het_type", "fin_type"]],
        how="inner", left_on="target_id", right_on="hetio_id"
    )
    .drop("hetio_id", axis=1)
    .rename(columns={
        "het_type": "target_htype",
        "fin_type": "target_ftype"
    })
    .drop_duplicates()
    .reset_index(drop=True)
)

het_etypes.shape

het_etypes.head()

### Group hetio edges by node types and count edges

h_etype_count = (het_etypes
    .groupby(["source_ftype", "target_ftype"])
    ["het_etype"]
    .value_counts()
    .to_frame("count")
    .reset_index()
)

h_etype_count

---

## Read SemmedDB edges

sedges = (pd
    .read_csv("../data/semmeddb/no_treats/semmeddb_no_treats_edges.tsv", sep='\t')
    .drop(["pmids", "n_pmids"], axis=1)
)

sedges.shape

sedges.head()

---

### Add in node types for SemmedDB

sem_etypes = (sedges
    .merge(
        snodes.drop("node_name", axis=1),
        how="inner", left_on="source_id", right_on="cui"
    )
    .drop("cui", axis=1)
    .rename(columns={"sem_type": "source_stype"})

    .merge(
        snodes.drop("node_name", axis=1),
        how="inner", left_on="target_id", right_on="cui"
    )
    .drop("cui", axis=1)
    .rename(columns={"sem_type": "target_stype"})
)

sem_etypes.shape

sem_etypes.head()

---

### Group SemmedDB edges by node type and count edges

s_etype_count = (sem_etypes
    .groupby(["source_stype", "target_stype"])
    ["edge_type"]
    .value_counts()
    .to_frame("count")
    .reset_index()
)

s_etype_count

---

## Combine all edges from both networks and manually decide which ones to retain

res = (h_etype_count
    .rename(columns={
        "source_ftype": "source_stype",
        "target_ftype": "target_stype",
        "het_etype": "edge_type"
    })
    .assign(orig = "h")
    .append(
        s_etype_count.assign(orig = "semmed")
    )
    .sort_values(
        by=["source_stype", "target_stype", "count"],
        ascending=[True, True, False]
    )
)

res

res.to_csv("../pipeline/manual_review/edge_merge.tsv", sep='\t', index=False)
