==> code/3_merge_hetionet_and_semmeddb.ipynb <==
# Create final merged network containing SemmedDB and Hetionet

2018-12-14

import pandas as pd

## Load in manual curation results for merging edges

Edge merging was performed in Excel from the `edge_merge.tsv` file.

res = (pd
    .read_csv("../pipeline/manual_review/edge_merge_final.tsv", sep='\t')
    .assign(changed = lambda df: df["edge_type"] != df["final_edge_type"])
)

res.shape

res.head()

### How many edge types were modified?

res["changed"].value_counts()

res.groupby(["changed", "orig"]).size()

Around half of the edges were untouched, and those which were changed were mostly hetionet edges (since we're trying to map onto semmeddb).

---

# Convert Hetionet edges to their final type

### Read Hetionet nodes and edges

hnodes = pd.read_csv("../pipeline/hetionet/final_hetionet_nodes_umls.tsv", sep='\t')

hnodes.shape

hnodes.head()

hedges = (pd
    .read_csv("../data/hetionet/raw/hetionet_edges.csv", sep=',')
    .rename(columns={
        ":START_ID": "source_id",
        ":END_ID": "target_id",
        ":TYPE": "het_etype"
    })
)

hedges.shape

hedges.head()

## Convert Hetionet edges

fhedges = (hedges
    # convert source node ids
    .merge(
        hnodes[["hetio_id", "final_id", "fin_type"]],
        how="inner", left_on="source_id", right_on="hetio_id"
    )
    .drop(["source_id", "hetio_id"], axis=1)
    .rename(columns={
        "final_id": "source_fid",
        "fin_type": "source_stype"
    })
     
    # convert target node ids
    .merge(
        hnodes[["hetio_id", "final_id", "fin_type"]],
        how="inner", left_on="target_id", right_on="hetio_id"
    )
    .drop(["target_id", "hetio_id"], axis=1)
    .rename(columns={
        "final_id": "target_fid",
        "fin_type": "target_stype"
    })
     
    .rename(columns={"het_etype": "edge_type"})

    # change edge types
    .merge(
        (res
            .query("final_edge_type != 'delete'")
            [["source_stype", "target_stype", "edge_type", "final_edge_type"]]
        ),
        how="inner", on=["source_stype", "target_stype", "edge_type"]
    )
    .drop("edge_type", axis=1)
     
    .assign(origin = "hetionet")
    .drop_duplicates()
    .reset_index(drop=True)
)

fhedges.shape

fhedges.head()

---

## Convert SemmedDB edges to their final types

### Read SemmedDB nodes

snodes = pd.read_csv("../data/semmeddb/no_treats/semmeddb_no_treats_nodes.tsv", sep='\t')

snodes.shape

snodes.head()

### Read SemmedDB edges

sedges = (pd
    .read_csv("../data/semmeddb/no_treats/semmeddb_no_treats_edges.tsv", sep='\t')
    .drop(["pmids", "n_pmids"], axis=1)
)

sedges.shape

sedges.head()

## Convert SemmedDB edges to their final types

fsedges = (sedges
    # add in source node type
    .merge(
        snodes.drop("node_name", axis=1),
        how="inner", left_on="source_id", right_on="node_id"
    )
    .drop("node_id", axis=1)
    .rename(columns={"node_type": "source_stype"})

    # add in target node type
    .merge(
        snodes.drop("node_name", axis=1),
        how="inner", left_on="target_id", right_on="node_id"
    )
    .drop("node_id", axis=1)
    .rename(columns={"node_type": "target_stype"})

    .rename(columns={
        "source_id": "source_fid",
        "target_id": "target_fid",
    })
           
    # change edge types
    .merge(
        (res
            .query("final_edge_type != 'delete'")
            [["source_stype", "target_stype", "edge_type", "final_edge_type"]]
        ),
        how="inner", on=["source_stype", "target_stype", "edge_type"]
    )
    .drop("edge_type", axis=1)
     
    .assign(origin = "semmeddb")
    .drop_duplicates()
    .reset_index(drop=True)
)

fsedges.shape

fsedges.head()

---

## Final results and write to file

edges_final = (fsedges
    .merge(
        fhedges, how="outer",
        on=["source_fid", "target_fid", "final_edge_type"]
    )
    .assign(
        source_stype = lambda df: df["source_stype_x"].fillna(df["source_stype_y"]),
        target_stype = lambda df: df["target_stype_x"].fillna(df["target_stype_y"]),
    )
    .drop([
        "source_stype_x", "source_stype_y",
        "target_stype_x", "target_stype_y"], axis=1
    )
    .fillna(False)
    .replace({
        "semmeddb": True,
        "hetionet": True
    })
    .rename(columns={
        "origin_x": "in_semmeddb",
        "origin_y": "in_hetionet"
    })
    .reset_index(drop=True)
)

edges_final.shape

edges_final.head()

## Merged network information

edges_final["source_fid"].nunique()

edges_final["target_fid"].nunique()

### Number of edges shared by both networks

edges_final.groupby(["in_semmeddb", "in_hetionet"]).size()

---

## Merged network node list

fin_nodes = (hnodes
    [["name", "final_id", "fin_type"]]
    .rename(columns={
        "name": "node_name",
        "final_id": "node_id",
        "fin_type": "node_type"
    })
    .assign(origin = "hetionet")
             
             
    .merge(
        snodes.assign(origin = "semmeddb"),
        how="outer", on="node_id"
    )
             
    .assign(
        node_type = lambda df: df["node_type_y"].fillna(df["node_type_x"]),
        node_name = lambda df: df["node_name_y"].fillna(df["node_name_x"])
    )
    .drop([
        "node_name_x", "node_name_y",
        "node_type_x", "node_type_y"], axis=1
    )
    .fillna(False)
    .replace({
        "semmeddb": True,
        "hetionet": True
    })
    .rename(columns={
        "origin_x": "in_hetionet",
        "origin_y": "in_semmeddb"
    })
    .drop_duplicates()
    .sort_values(["node_id", "node_type"])
    .reset_index(drop=True)
)

fin_nodes.shape

fin_nodes.head()

### Merged network node information

fin_nodes["node_id"].nunique()

fin_nodes.groupby(["in_hetionet", "in_semmeddb"]).size()

fin_nodes["node_type"].value_counts()

---

## Save merged network to disk

fin_nodes.to_csv("../pipeline/merged_network/merged_hetionet_and_semmeddb_nodes.tsv", sep='\t', index=False)

edges_final.to_csv("../pipeline/merged_network/merged_hetionet_and_semmeddb_edges.tsv", sep='\t', index=False)
